+++ blaze-client/src/main/scala/org/http4s/client/blaze/Http1Connection.scala
@@ -390,10 +390,7 @@ private object Http1Connection {
 
   private def encodeRequestLine[F[_]](req: Request[F], writer: Writer): writer.type = {
     val uri = req.uri
-    writer << req.method << ' ' << uri.copy(
-      scheme = None,
-      authority = None,
-      fragment = None) << ' ' << req.httpVersion << "\r\n"
+    writer << req.method << ' ' << uri.toOriginForm << ' ' << req.httpVersion << "\r\n"
     if (getHttpMinor(req) == 1 && Host
         .from(req.headers)
         .isEmpty) { // need to add the host header for HTTP/1.1