+++ core/src/main/scala/org/http4s/Media.scala
@@ -16,7 +16,7 @@
 
 package org.http4s
 
-import cats.MonadError
+import cats.MonadThrow
 import cats.syntax.all._
 import fs2.{RaiseThrowable, Stream}
 import fs2.text.utf8Decode
@@ -78,7 +78,7 @@ trait Media[F[_]] {
     * @tparam A type of the result
     * @return the effect which will generate the A
     */
-  final def as[A](implicit F: MonadError[F, Throwable], decoder: EntityDecoder[F, A]): F[A] =
+  final def as[A](implicit F: MonadThrow[F], decoder: EntityDecoder[F, A]): F[A] =
     // n.b. this will be better with redeem in Cats-2.0
     attemptAs.leftWiden[Throwable].rethrowT
 }