+++ ember-core/src/test/scala/org/http4s/ember/core/ParsingSpec.scala
@@ -257,7 +257,6 @@ class ParsingSpec extends Specification with CatsIO {
           for {
             body <- resp.body.through(text.utf8Decode).compile.string
             trailers <- resp.trailerHeaders
-            _ = println(trailers)
           } yield (body must beEqualTo("MozillaDeveloperNetwork")).and(
             trailers.get(Expires) must beSome
           )
@@ -319,8 +318,6 @@ class ParsingSpec extends Specification with CatsIO {
         .compile
         .lastOrError
         .unsafeRunSync()
-
-      println(headers)
 
       headers.toList must_=== List(
         Header("Content-Type", "text/plain"),