+++ ember-server/src/main/scala/org/http4s/ember/server/EmberServerBuilder.scala
@@ -19,17 +19,18 @@ package org.http4s.ember.server
 import cats._
 import cats.syntax.all._
 import cats.effect._
-import fs2.concurrent._
+import cats.effect.concurrent._
 import fs2.io.tcp.SocketGroup
 import fs2.io.tcp.SocketOptionMapping
 import fs2.io.tls._
 import org.http4s._
 import org.http4s.server.Server
+
 import scala.concurrent.duration._
 import java.net.InetSocketAddress
 import _root_.io.chrisdavenport.log4cats.Logger
 import _root_.io.chrisdavenport.log4cats.slf4j.Slf4jLogger
-import org.http4s.ember.server.internal.ServerHelpers
+import org.http4s.ember.server.internal.{ServerHelpers, Shutdown}
 
 final class EmberServerBuilder[F[_]: Concurrent: Timer: ContextShift] private (
     val host: String,
@@ -38,50 +39,17 @@ final class EmberServerBuilder[F[_]: Concurrent: Timer: ContextShift] private (
     private val blockerOpt: Option[Blocker],
     private val tlsInfoOpt: Option[(TLSContext, TLSParameters)],
     private val sgOpt: Option[SocketGroup],
-    private val onError: Throwable => Response[F],
+    private val errorHandler: Throwable => F[Response[F]],
     private val onWriteFailure: (Option[Request[F]], Response[F], Throwable) => F[Unit],
     val maxConcurrency: Int,
     val receiveBufferSize: Int,
     val maxHeaderSize: Int,
     val requestHeaderReceiveTimeout: Duration,
     val idleTimeout: Duration,
+    val shutdownTimeout: Duration,
     val additionalSocketOptions: List[SocketOptionMapping[_]],
     private val logger: Logger[F]
 ) { self =>
-
-  @deprecated("Kept for binary compatibility", "0.21.7")
-  private[EmberServerBuilder] def this(
-      host: String,
-      port: Int,
-      httpApp: HttpApp[F],
-      blockerOpt: Option[Blocker],
-      tlsInfoOpt: Option[(TLSContext, TLSParameters)],
-      sgOpt: Option[SocketGroup],
-      onError: Throwable => Response[F],
-      onWriteFailure: (Option[Request[F]], Response[F], Throwable) => F[Unit],
-      maxConcurrency: Int,
-      receiveBufferSize: Int,
-      maxHeaderSize: Int,
-      requestHeaderReceiveTimeout: Duration,
-      additionalSocketOptions: List[SocketOptionMapping[_]],
-      logger: Logger[F]) =
-    this(
-      host = host,
-      port = port,
-      httpApp = httpApp,
-      blockerOpt = blockerOpt,
-      tlsInfoOpt = tlsInfoOpt,
-      sgOpt = sgOpt,
-      onError = onError,
-      onWriteFailure = onWriteFailure,
-      maxConcurrency = maxConcurrency,
-      receiveBufferSize = receiveBufferSize,
-      maxHeaderSize = maxHeaderSize,
-      requestHeaderReceiveTimeout = requestHeaderReceiveTimeout,
-      idleTimeout = EmberServerBuilder.Defaults.idleTimeout,
-      additionalSocketOptions = additionalSocketOptions,
-      logger = logger
-    )
 
   private def copy(
       host: String = self.host,
@@ -90,13 +58,14 @@ final class EmberServerBuilder[F[_]: Concurrent: Timer: ContextShift] private (
       blockerOpt: Option[Blocker] = self.blockerOpt,
       tlsInfoOpt: Option[(TLSContext, TLSParameters)] = self.tlsInfoOpt,
       sgOpt: Option[SocketGroup] = self.sgOpt,
-      onError: Throwable => Response[F] = self.onError,
+      errorHandler: Throwable => F[Response[F]] = self.errorHandler,
       onWriteFailure: (Option[Request[F]], Response[F], Throwable) => F[Unit] = self.onWriteFailure,
       maxConcurrency: Int = self.maxConcurrency,
       receiveBufferSize: Int = self.receiveBufferSize,
       maxHeaderSize: Int = self.maxHeaderSize,
       requestHeaderReceiveTimeout: Duration = self.requestHeaderReceiveTimeout,
       idleTimeout: Duration = self.idleTimeout,
+      shutdownTimeout: Duration = self.shutdownTimeout,
       additionalSocketOptions: List[SocketOptionMapping[_]] = self.additionalSocketOptions,
       logger: Logger[F] = self.logger
   ): EmberServerBuilder[F] =
@@ -107,13 +76,14 @@ final class EmberServerBuilder[F[_]: Concurrent: Timer: ContextShift] private (
       blockerOpt = blockerOpt,
       tlsInfoOpt = tlsInfoOpt,
       sgOpt = sgOpt,
-      onError = onError,
+      errorHandler = errorHandler,
       onWriteFailure = onWriteFailure,
       maxConcurrency = maxConcurrency,
       receiveBufferSize = receiveBufferSize,
       maxHeaderSize = maxHeaderSize,
       requestHeaderReceiveTimeout = requestHeaderReceiveTimeout,
       idleTimeout = idleTimeout,
+      shutdownTimeout = shutdownTimeout,
       additionalSocketOptions = additionalSocketOptions,
       logger = logger
     )
@@ -136,7 +106,16 @@ final class EmberServerBuilder[F[_]: Concurrent: Timer: ContextShift] private (
   def withIdleTimeout(idleTimeout: Duration) =
     copy(idleTimeout = idleTimeout)
 
-  def withOnError(onError: Throwable => Response[F]) = copy(onError = onError)
+  def withShutdownTimeout(shutdownTimeout: Duration) =
+    copy(shutdownTimeout = shutdownTimeout)
+
+  @deprecated("0.21.17", "Use withErrorHandler - Do not allow the F to fail")
+  def withOnError(onError: Throwable => Response[F]) =
+    withErrorHandler({ case e => onError(e).pure[F] })
+
+  def withErrorHandler(errorHandler: PartialFunction[Throwable, F[Response[F]]]) =
+    copy(errorHandler = errorHandler)
+
   def withOnWriteFailure(onWriteFailure: (Option[Request[F]], Response[F], Throwable) => F[Unit]) =
     copy(onWriteFailure = onWriteFailure)
   def withMaxConcurrency(maxConcurrency: Int) = copy(maxConcurrency = maxConcurrency)
@@ -148,10 +127,11 @@ final class EmberServerBuilder[F[_]: Concurrent: Timer: ContextShift] private (
 
   def build: Resource[F, Server[F]] =
     for {
+      bindAddress <- Resource.liftF(Sync[F].delay(new InetSocketAddress(host, port)))
       blocker <- blockerOpt.fold(Blocker[F])(_.pure[Resource[F, *]])
       sg <- sgOpt.fold(SocketGroup[F](blocker))(_.pure[Resource[F, *]])
-      bindAddress <- Resource.liftF(Sync[F].delay(new InetSocketAddress(host, port)))
-      shutdownSignal <- Resource.liftF(SignallingRef[F, Boolean](false))
+      ready <- Resource.liftF(Deferred[F, Either[Throwable, Unit]])
+      shutdown <- Resource.liftF(Shutdown[F](shutdownTimeout))
       _ <- Concurrent[F].background(
         ServerHelpers
           .server(
@@ -159,9 +139,10 @@ final class EmberServerBuilder[F[_]: Concurrent: Timer: ContextShift] private (
             httpApp,
             sg,
             tlsInfoOpt,
-            onError,
+            ready,
+            shutdown,
+            errorHandler,
             onWriteFailure,
-            shutdownSignal.some,
             maxConcurrency,
             receiveBufferSize,
             maxHeaderSize,
@@ -173,7 +154,8 @@ final class EmberServerBuilder[F[_]: Concurrent: Timer: ContextShift] private (
           .compile
           .drain
       )
-      _ <- Resource.make(Applicative[F].unit)(_ => shutdownSignal.set(true))
+      _ <- Resource.make(Applicative[F].unit)(_ => shutdown.await)
+      _ <- Resource.liftF(ready.get.rethrow)
     } yield new Server[F] {
       def address: InetSocketAddress = bindAddress
       def isSecure: Boolean = tlsInfoOpt.isDefined
@@ -189,13 +171,14 @@ object EmberServerBuilder {
       blockerOpt = None,
       tlsInfoOpt = None,
       sgOpt = None,
-      onError = Defaults.onError[F],
+      errorHandler = Defaults.errorHandler[F],
       onWriteFailure = Defaults.onWriteFailure[F],
       maxConcurrency = Defaults.maxConcurrency,
       receiveBufferSize = Defaults.receiveBufferSize,
       maxHeaderSize = Defaults.maxHeaderSize,
       requestHeaderReceiveTimeout = Defaults.requestHeaderReceiveTimeout,
       idleTimeout = Defaults.idleTimeout,
+      shutdownTimeout = Defaults.shutdownTimeout,
       additionalSocketOptions = Defaults.additionalSocketOptions,
       logger = Slf4jLogger.getLogger[F]
     )
@@ -205,9 +188,20 @@ object EmberServerBuilder {
     val port: Int = server.defaults.HttpPort
 
     def httpApp[F[_]: Applicative]: HttpApp[F] = HttpApp.notFound[F]
-    def onError[F[_]]: Throwable => Response[F] = { (_: Throwable) =>
-      Response[F](Status.InternalServerError)
+
+    private val serverFailure =
+      Response(Status.InternalServerError).putHeaders(org.http4s.headers.`Content-Length`.zero)
+    // Effectful Handler - Perhaps a Logger
+    // Will only arrive at this code if your HttpApp fails or the request receiving fails for some reason
+    def errorHandler[F[_]: Applicative]: Throwable => F[Response[F]] = { case (_: Throwable) =>
+      serverFailure.covary[F].pure[F]
     }
+
+    @deprecated("0.21.17", "Use errorHandler, default fallback of failure InternalServerFailure")
+    def onError[F[_]]: Throwable => Response[F] = { (_: Throwable) =>
+      serverFailure.covary[F]
+    }
+
     def onWriteFailure[F[_]: Applicative]
         : (Option[Request[F]], Response[F], Throwable) => F[Unit] = {
       case _: (Option[Request[F]], Response[F], Throwable) => Applicative[F].unit
@@ -217,6 +211,7 @@ object EmberServerBuilder {
     val maxHeaderSize: Int = server.defaults.MaxHeadersSize
     val requestHeaderReceiveTimeout: Duration = 5.seconds
     val idleTimeout: Duration = server.defaults.IdleTimeout
+    val shutdownTimeout: Duration = server.defaults.ShutdownTimeout
     val additionalSocketOptions = List.empty[SocketOptionMapping[_]]
   }
 }