+++ servlet/src/test/scala/org/http4s/servlet/BlockingHttp4sServletSuite.scala
@@ -47,8 +47,7 @@ class BlockingHttp4sServletSuite extends Http4sSuite {
     }
     .orNotFound
 
-  def servletServer: FunFixture[Int] =
-    ResourceFixture[Int](serverPortR)
+  val servletServer = ResourceFixture[Int](serverPortR)
 
   def get(serverPort: Int, path: String): IO[String] =
     testBlocker.delay[IO, String](